+++ CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
 ### Latest changes
 
 * VFS-9614 Added archivisation audit log
+* VFS-9801 Added number of all processed (scanned) files in transfer
+* VFS-9729 Improved showing storage locations for file in data distribution modal and file info modal
 * VFS-9712 Added detailed view of upload errors
 * VFS-9513 Added file path tooltip to file event in QoS audit log
 * VFS-9623 Increased speed of SASS compilation and fixed its unnecessary recompilation